feat(execution): integrate execution_contracts for backend functionality and refactor related components
