+++ execution/src/lib.rs
@@ -1,12 +1,19 @@
-mod backend;
+macro_rules! for_each_backend {
+    ($macro:ident) => {
+        $macro!(Cpu => execution_cpu::CpuBundle);
+    };
+}
+
 mod capability;
-mod contracts;
 mod kernel;
 mod storage;
 mod tag;
 
 pub use capability::Capability;
-pub use contracts::*;
+pub use execution_contracts::{
+    BackendBundle, ExecutionCapability, ExecutionKernelArgs, ExecutionKernelContext,
+    ExecutionKernelLauncher, ExecutionKernelMetadata, ExecutionStorage, ExecutionStorageContext,
+};
 pub use execution_cpu::{
     CpuBuffer, CpuBundle, CpuCapability, CpuExecution, CpuKernelArgs, CpuKernelContext,
     CpuKernelEntrypoint, CpuKernelFn, CpuKernelLaunchConfig, CpuKernelLaunchError,
