+++ execution_contracts/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
-name = "execution_cpu"
+name = "execution_contracts"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-execution_contracts = { path = "../execution_contracts" }
 schema = { path = "../schema" }
