+++ execution_contracts/src/lib.rs
@@ -1,0 +1,8 @@
+mod bundle;
+mod contracts;
+
+pub use bundle::BackendBundle;
+pub use contracts::{
+    ExecutionCapability, ExecutionKernelArgs, ExecutionKernelContext, ExecutionKernelLauncher,
+    ExecutionKernelMetadata, ExecutionStorage, ExecutionStorageContext,
+};
